Update 12.03.23
The game is pretty much done!

Just need to add some details, like a button in the end of the story  to go to the credits and some UI sounds whenb the player clicks on it.

Need to learn how to create my own game page on itch.io and post it to make it public to the people.